trail_project/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
openai>=1.30.0
jinja2>=3.1.0
pyyaml>=6.0
xgboost>=2.0.0
statsmodels>=0.14.0
biogeme>=3.2.13
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
python-dotenv>=1.0.0
pyarrow>=14.0.0
pyodbc>=5.0.0
pypdf>=4.0.0
